[cmd] `set` command: added missing issue meta fields
So far `set` command was only implemented for `state`.

Not it is also implemented for the other editable fields:
`title`, `type`, `assignee` and `labels`.

For `label`: The specified labels will be added, if not already
exists; so we don't overwrite existing labels. -- To remove
existing labels a separate command will be added in the future.

Signed-off-by: Timon Burkard <burkard.timon@gmail.com>