[cmd/show] improved representation by using external editor
A temporary markdown file is created, which contains all the info
about an issue:

```md
<!-- READ-ONLY VIEW -->

# Issue #<id> -- <title>

## Meta Data

| **field**    | **value** |
| ------------ | --------- |
| **id**       | <id> |
| **title**    | <title> |
| **state**    | <state> |
| **type**     | <type> |
| **labels**   | <label> |
| **assignee** | <assignee> |
| **created**  | <timestamp> |
| **updated**  | <timestamp> |

## Description

### Description

### Repro Steps

### Expected Behavior

### System Info
```

This temporary markdown file is then opened with the configured
external editor.

Also all the issue `attachments/` are copied, so that the relative
links still work. The user can therefore nicely view the markdown
preview, e.g., with vscode.

Signed-off-by: Timon Burkard <burkard.timon@gmail.com>