[meta] limit options for `assignee` field by introducing `users.yaml`
The available assignees are configurable in `users.yaml:users:id`

The default, set at issue creation, is still empty (`''`). Which is
also a valid option that can be set.

At `git issue init` the `config/users-default.yaml` is copied to
`.gitissues/users.yaml`, which can then be edited by the user.

Signed-off-by: Timon Burkard <burkard.timon@gmail.com>