[config] added possibility to change ID generation strategy
With the ID generation so far, which uses sequential numbers by
scanning through the `gitissues/issues/` directory, find the max
and then take max + 1, for teams this gives merge conflict when
two users create a new issue at the same time (before merging).
 --> `sequential`

Therefore, a new ID generation strategy was introduced for teams:
 --> `timestamp`

This takes the current timestamp in seconds since 2025-01-01. This
significantly reduces the chance of merge conflicts.

Signed-off-by: Timon Burkard <burkard.timon@gmail.com>