[meta] allowed priority to be set to empty
Previously, it was not possible for an issue to have empty priority.

But, if someone doesn't want to use priorities, it would be nice if
he could configure default priority to be empty.
 --> This is now possible: `config.yaml:priority_default`

Or, if for an the priority for an issue, has not yet been defined,
it might be useful to set it to empty.
--> This is now possible: `git issue set <id> --priority ''`

Signed-off-by: Timon Burkard <burkard.timon@gmail.com>